Remove confusing overlapping ICompats from MCIL & MMH
