+++ src/main/java/metadev/digital/metacustomitemslib/compatibility/addons/CMICompat.java
@@ -4,7 +4,7 @@ import com.Zrips.CMI.CMI;
 import com.Zrips.CMI.Modules.Holograms.HologramManager;
 import metadev.digital.metacustomitemslib.compatibility.Feature;
 import metadev.digital.metacustomitemslib.compatibility.FeatureList;
-import metadev.digital.metacustomitemslib.compatibility.ICompat;
+import metadev.digital.metacustomitemslib.compatibility.ICustomItemsLibCompat;
 import metadev.digital.metacustomitemslib.compatibility.IFeatureHolder;
 import metadev.digital.metacustomitemslib.compatibility.enums.BoundIdentifierEnum;
 import metadev.digital.metacustomitemslib.compatibility.enums.SupportedPluginEntities;
@@ -16,7 +16,7 @@ import metadev.digital.metacustomitemslib.Core;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.Plugin;
 
-public class CMICompat implements ICompat, IFeatureHolder {
+public class CMICompat implements ICustomItemsLibCompat, IFeatureHolder {
 
 	// ****** Standard ******
 	private static Plugin compatPlugin;
